requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
pytorch-lightning
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
gdown
flask
Werkzeug

# Logging -------------------------------------
tensorboard>=2.4.1
comet-ml

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

thop  # FLOPs computation
#added
torchvision
onnx
onnxruntime
openvino-dev